third_party/WebKit/LayoutTests/http/tests/inspector/bindings/suspendtarget-bindings.html

<head>
<script src="../inspector-test.js"></script>
<script src="./bindings-test.js"></script>

<script>

async function test()
{
    InspectorTest.markStep('initialWorkspace');
    var snapshot = InspectorTest.dumpWorkspace();

    InspectorTest.markStep('createIframesAndWaitForSourceMaps');
    await Promise.all([
        InspectorTest.attachFrame('frame1', './resources/sourcemap-frame.html', '_test_create-iframe1.js'),
        InspectorTest.attachFrame('frame2', './resources/sourcemap-frame.html', '_test_create-iframe2.js'),
        InspectorTest.waitForSourceMap('sourcemap-script.js.map'),
        InspectorTest.waitForSourceMap('sourcemap-style.css.map')
    ]);
    snapshot = InspectorTest.dumpWorkspace(snapshot);

    InspectorTest.markStep('Suspending targets.');
    await SDK.targetManager.suspendAllTargets();
    snapshot = InspectorTest.dumpWorkspace(snapshot);

    InspectorTest.markStep('detachFrame');
    await InspectorTest.detachFrame('frame1', '_test_detachFrame.js');
    snapshot = InspectorTest.dumpWorkspace(snapshot);

    InspectorTest.markStep('Resuming targets.');
    await Promise.all([
        SDK.targetManager.resumeAllTargets(),
        InspectorTest.waitForSourceMap('sourcemap-script.js.map'),
        InspectorTest.waitForSourceMap('sourcemap-style.css.map')
    ]);
    snapshot = InspectorTest.dumpWorkspace(snapshot);

    InspectorTest.completeTest();
}

</script>
</head>
<body onload="runTest()">
<p>
Verify that bindings handle target suspension as expected.
</p>
</body>
